kubespray_orchestration_script/requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cloudshell-orch-core==3.4.0.204
cryptography==36.0.1
idna==3.3
paramiko==2.9.1
pycparser==2.21
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
PyYAML==6.0
